% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/early_learning.R
\name{early_learning}
\alias{early_learning}
\title{obtained early learning scores}
\usage{
early_learning(input, min.responses = 1, ...)
}
\arguments{
\item{input}{A data.frame with columns years and items in lex_kidsights (prefix = AA,BB,CC, or DD)..}

\item{min.responses}{Minimum number of responses to items for scoring an individual. Defaults to 5.}

\item{...}{Other arguments passed to mirt::fscores}
}
\value{
A data.frame (or list) of Kidsights scores
}
\description{
obtained early learning scores
}
