% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscores.R
\name{fscores}
\alias{fscores}
\title{Score Kidsights}
\usage{
fscores(input, est_hyperpriors = T, plausible_values = 0)
}
\arguments{
\item{input}{A data.frame with columns years and items in lex_kidsights (prefix = AA,BB,CC, or DD)..}

\item{est_hyperpriors}{Defaults to True. Estimates the hyperpriors meand and variances based on available data.}

\item{plausible_values}{Defaults to 0, in which case the EAP estimate is returned.}
}
\value{
A data.frame (or list) of EAP scores (or plausible values).
}
\description{
Score Kidsights
}
