% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kidsights_scores.R
\name{kidsights_scores}
\alias{kidsights_scores}
\title{Score Kidsights}
\usage{
kidsights_scores(input, version = "version 2.0", min.responses = 1, ...)
}
\arguments{
\item{input}{A data.frame with columns years and items in lex_kidsights (prefix = AA,BB,CC, or DD)..}

\item{version}{Defaults to the most current version ("version 2.0"). Other options include "version 1.0".}

\item{min.responses}{Minimum number of responses to items for scoring an individual. Defaults to 5.}

\item{...}{Other arguments passed to mirt::fscores}
}
\value{
A data.frame (or list) of Kidsights scores
}
\description{
Score Kidsights
}
